Add separate array pointer types, indexing
The pointer offset operator `++` has been removed.  In its place we have
a new pair of pointer types, `refs T` and `muts T`, which represent
pointers that point to an array of values.  Under the hood these are
just pointers, but they help type safety by clarifying which pointers
are just used for indirection and which are used for arrays.  Array
pointers can be indexed using the standard `a[i]` subscripting syntax.