Infer integer literal types when possible
This is a simple form of type inference that only looks at the immediate
context of the integer literal to guess its type (with the exception of
unary minus, which is treated specially to allow the inference to see
through it).  This hopefully prevents the confusing error messages that
often result from highly non-local type inference, while ideally
providing enough inference for most use cases (such as array indexing).